move helpers to /lib
