+++ src/desktop/apps/article/components/InfiniteScrollArticle.tsx
@@ -11,7 +11,7 @@ import {
 import { articlesQuery } from "desktop/apps/article/queries/articles"
 import { ArticleData } from "@artsy/reaction/dist/Components/Publishing/Typings"
 import { shouldAdRender } from "desktop/apps/article/helpers"
-import { openAuthModal } from "desktop/apps/authentication/helpers"
+import { openAuthModal } from "desktop/lib/openAuthModal"
 
 const FETCH_TOP_OFFSET = 200
