+++ src/desktop/apps/how_auctions_work/client/index.coffee
@@ -1,5 +1,5 @@
 multiPageView = require '../../../components/multi_page/index.coffee'
-{ openAuthModal } = require '../../authentication/helpers'
+{ openAuthModal } = require '../../../lib/openAuthModal'
 { ModalType } = require "@artsy/reaction/dist/Components/Authentication/Types"
 
 module.exports.init = ->
