+++ src/desktop/components/artwork_item/views/sale_artwork.coffee
@@ -2,7 +2,7 @@ _ = require 'underscore'
 sd = require('sharify').data
 Backbone = require 'backbone'
 ContactPartnerView = require '../../contact/contact_partner.coffee'
-{ openAuthModal } = require '../../../apps/authentication/helpers'
+{ openAuthModal } = require '../../../lib/openAuthModal'
 { ModalType } = require "@artsy/reaction/dist/Components/Authentication/Types"
 
 module.exports = class SaleArtworkView extends Backbone.View
