+++ src/desktop/components/artwork_save/view.coffee
@@ -1,6 +1,6 @@
 Backbone = require 'backbone'
 analyticsHooks = require '../../lib/analytics_hooks.coffee'
-{ openAuthModal } = require '../../apps/authentication/helpers'
+{ openAuthModal } = require '../../lib/openAuthModal'
 { ModalType } = require "@artsy/reaction/dist/Components/Authentication/Types"
 
 module.exports = class ArtworkSaveView extends Backbone.View
