+++ src/desktop/components/fair_layout/client/header.coffee
@@ -3,7 +3,7 @@ Backbone = require 'backbone'
 sd = require('sharify').data
 Profile = require '../../../models/profile.coffee'
 Fair = require '../../../models/fair.coffee'
-{ openAuthModal } = require '../../../apps/authentication/helpers'
+{ openAuthModal } = require '../../../lib/openAuthModal'
 { ModalType } = require "@artsy/reaction/dist/Components/Authentication/Types"
 CurrentUser = require '../../../models/current_user.coffee'
 FlashMessage = require '../../flash/index.coffee'
